[WIP] Fixed scheduling work
Unit tests still have issues
